Stylesheets cleanup. Refactor cleanup.
