+++ src/sql/workbench/browser/modal/media/modal.css
@@ -24,112 +24,12 @@
 	height: 480px;
 }
 
-.modal.callout-dialog {
-	background-color: transparent;
-}
-.modal.callout-dialog .modal-dialog {
-	border-radius: 2px;
-	box-shadow: 0px 3px 8px rgba(var(--foreground));
-	max-height: 300px;
-	position: absolute;
-}
-
-.modal.callout-dialog .modal-content .insert-image .flex-container {
-	display: flex;
-}
-.modal.callout-dialog .modal-content .insert-image .flex-container > div {
-	flex: 1;
-}
-.modal.callout-dialog .modal-content p {
-	margin: 0;
-}
-.modal.callout-dialog .modal-content .button-icon {
-	cursor: pointer;
-	margin-left: 10px;
-}
-.modal.callout-dialog .modal-content .insert-image .monaco-inputbox {
-	min-width: 380px;
-}
-.modal.callout-dialog .modal-content .row  {
-	margin-bottom: 16px;
-}
-.modal.callout-dialog .modal-content .radio-group input {
-	margin-right: 8px;
-}
-.modal.callout-dialog .modal-content .radio-group span {
-	margin-right: 15px;
-}
-
-.hc-black .modal.callout-dialog .modal-dialog {
-	box-shadow: none;
-}
-
-/* Correct the arrow appearance for HC theme */
-.callout-arrow:before {
-	border-width: 1px;
-	border-style: solid;
-	border-color:
-		transparent
-		transparent
-		var(--bodybackground)
-		var(--bodybackground);
-	box-shadow: -3px 3px 3px 0 rgba(var(--foreground));
-	content: '';
-	display: block;
-	height: 0;
-	position: absolute;
-	width: 0;
-}
-.callout-arrow.from-below:before {
-	border-width: 0.5em;
-	left: 2em;
-	top: -0.2em;
-	transform: rotate(135deg);
-}
-.callout-arrow.from-left:before {
-	background-color: var(--bodybackground);
-	height: 26px;
-	right: -13px;
-	top: 26px;
-	transform: rotate(-135deg);
-	width: 26px;
-}
-
-.hc-black .callout-arrow:before {
-	background-color: var(--bodybackground);
-	border-color:
-		transparent
-		transparent
-		var(--border)
-		var(--border);
-	border-width: 0.1em;
-	box-shadow: none;
-	height: 0.8em;
-	width: 0.8em;
-}
-.hc-black .callout-arrow.from-below:before {
-	top: -0.4em;
-}
-.hc-black .callout-arrow.from-left:before {
-	height: 2em;
-	right: -1.2em;
-	width: 2em;
-}
-
-
 .modal .modal-header {
 	padding: 15px;
 }
 
-.modal.callout-dialog .modal-header {
-	padding: 18px 24px 8px 24px;
-}
-
 .modal .modal-footer {
 	padding: 15px;
-}
-.modal.callout-dialog .modal-footer {
-	padding: 15px 24px 15px 24px;
 }
 
 .modal .codicon.in-progress {
@@ -187,17 +87,6 @@
 .modal.flyout-dialog .modal-body-and-footer {
 	flex: 1 1;
 	overflow: hidden;
-}
-
-.modal.callout-dialog .modal-body {
-	padding: 8px 24px;
-}
-
-.modal.callout-dialog.compact .modal-header {
-	padding: 16px 24px 4px 24px;
-}
-.modal.callout-dialog.compact .modal-body {
-	padding: 4px 24px 16px 24px;
 }
 
 /* modl body content style(excluding dialogErrorMessage section) for angulr component dialog */
