+++ src/sql/workbench/contrib/notebook/browser/calloutDialog/common/constants.ts
@@ -1,0 +1,9 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the Source EULA. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+import { localize } from 'vs/nls';
+
+// Localized texts
+export const insertButtonText = localize('callout.insertButton', "Insert");
+export const cancelButtonText = localize('callout.cancelButton', "Cancel");
