+++ src/sql/workbench/contrib/notebook/browser/calloutDialog/media/linkCalloutDialog.css
@@ -1,0 +1,5 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the Source EULA. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
